Fixation is the way to death. Fluidity is the way to life.
