+++ DAO/project/contracts/DAO-contract/tests/functions/core/mod.rs
@@ -1,0 +1,7 @@
+mod constructor;
+mod create_proposal;
+mod deposit;
+mod execute;
+mod unlock_votes;
+mod vote;
+mod withdraw;
