+++ DAO/project/contracts/DAO-contract/tests/functions/info/mod.rs
@@ -1,0 +1,6 @@
+mod balance;
+mod governance_token_id;
+mod proposal;
+mod proposal_count;
+mod user_balance;
+mod user_votes;
